updated ISR's

+++ simptris.h
@@ -2,6 +2,6 @@
 
 void SlidePiece(int ID, int direction);
 void RotatePiece(int ID, int direction);
-void SeedSimptris(long seed);
+void SeedSimptris(int seed);
 void StartSimptris(void);
